add map with geotagged entries to Dashboard
This commit adds a SimpleMap component to Dashboard, below the
rest of the user interface. The component takes as props the
"this.state.geotaggedEntries" array, and maps over those entries,
creating a Pin component for each entry. inline comments explain the
Pin component if it's not clear how the props work.

The Pin components have a "props.$hover" that decides their style,
you can consult the Google-map-react documentation to find out more